+++ packages/sdk/src/contract-functions/getLBTCMintingFee/getLBTCMintingFee.tsx
@@ -3,8 +3,9 @@ import { CommonParameters } from '../../common/parameters';
 import { determineEnv } from '../../utils/env';
 import { fromSatoshi } from '../../utils/satoshi';
 import { makePublicClient } from '../../clients/public-client';
-import { getTokenContractInfo } from '../../tokens/tokens';
+import { getTokenContractInfo, isSTLBTCAbi } from '../../tokens/tokens';
 import { Token } from '../../tokens/token-addresses';
+import ASSET_ROUTER_ABI from '../../tokens/abi/ASSET_ROUTER_ABI';
 
 /**
  * Gets LBTC minting fee amount.
@@ -24,7 +25,7 @@ export async function getLBTCMintingFee(
 
 /** Gets LBTC burn commission (fee) amount. */
 export async function getLBTCBurningFee(props: CommonParameters) {
-  return getBurningFee({ token: Token.LBTC, ...props });
+  return getRedeemFee({ token: Token.LBTC, ...props });
 }
 
 export async function getMintingFee({
@@ -33,44 +34,110 @@ export async function getMintingFee({
   rpcUrl,
   env,
 }: { token: Token } & CommonParameters) {
-  if (![Token.LBTC, Token.BTCK].includes(token)) {
+  if (![Token.LBTC, Token.BTCK, Token.NativeLBTC].includes(token)) {
     throw new Error(`Unsupported token: ${token}`);
   }
 
   const environment = env || determineEnv(chainId);
 
   const publicClient = makePublicClient({ chainId, rpcUrl, env: environment });
-  const lbtcContract = getTokenContractInfo(token, chainId, environment);
+  const tokenContract = getTokenContractInfo(token, chainId, environment);
+  const tokenContractAbi = tokenContract.abi;
 
-  const rawFeeValue = await publicClient.readContract({
-    abi: lbtcContract.abi,
-    address: lbtcContract.address,
-    functionName: 'getMintFee',
-  });
+  let rawFeeValue = 0n;
+
+  if (isSTLBTCAbi(tokenContractAbi) || token === Token.NativeLBTC) {
+    const assetRouterAddress = await publicClient.readContract({
+      abi: tokenContractAbi,
+      address: tokenContract.address,
+      functionName: 'getAssetRouter',
+    });
+
+    const assetRouter = {
+      abi: ASSET_ROUTER_ABI,
+      address: assetRouterAddress,
+    };
+
+    rawFeeValue = await publicClient.readContract({
+      abi: assetRouter.abi,
+      address: assetRouter.address,
+      functionName: 'maxMintCommission',
+    });
+  } else {
+    rawFeeValue = await publicClient.readContract({
+      abi: tokenContract.abi,
+      address: tokenContract.address,
+      functionName: 'getMintFee',
+    });
+  }
 
   return fromSatoshi(String(rawFeeValue));
 }
 
-export async function getBurningFee({
+export async function getRedeemFee({
   token,
   chainId,
   rpcUrl,
   env,
 }: { token: Token } & CommonParameters) {
-  if (![Token.LBTC, Token.BTCK].includes(token)) {
+  if (![Token.LBTC, Token.BTCK, Token.NativeLBTC].includes(token)) {
     throw new Error(`Unsupported token: ${token}`);
   }
 
   const environment = env || determineEnv(chainId);
 
   const publicClient = makePublicClient({ chainId, rpcUrl, env: environment });
-  const lbtcContract = getTokenContractInfo(token, chainId, environment);
+  const tokenContract = getTokenContractInfo(token, chainId, environment);
 
-  const rawFeeValue = await publicClient.readContract({
-    abi: lbtcContract.abi,
-    address: lbtcContract.address,
-    functionName: 'getBurnCommission',
-  });
+  let rawFeeValue = 0n;
+  if (
+    (token === Token.LBTC && isSTLBTCAbi(tokenContract.abi)) ||
+    token === Token.NativeLBTC
+  ) {
+    const nativeTokenContract = getTokenContractInfo(
+      Token.NativeLBTC,
+      chainId,
+      environment,
+    );
+
+    const assetRouterAddress = await publicClient.readContract({
+      abi: tokenContract.abi,
+      address: tokenContract.address,
+      functionName: 'getAssetRouter',
+    });
+
+    const assetRouter = {
+      abi: ASSET_ROUTER_ABI,
+      address: assetRouterAddress,
+    };
+
+    const toNativeCommissionValue = await publicClient.readContract({
+      abi: assetRouter.abi,
+      address: assetRouter.address,
+      functionName: 'toNativeCommission',
+    });
+
+    const [redeemFeeValue /* redeemForBtcMinAmountValue, isRedeemEnabled */] =
+      await publicClient.readContract({
+        abi: assetRouter.abi,
+        address: assetRouter.address,
+        functionName: 'tokenConfig',
+        args: [nativeTokenContract.address],
+      });
+
+    if (token === Token.LBTC) {
+      rawFeeValue = toNativeCommissionValue + redeemFeeValue;
+    } else {
+      rawFeeValue = redeemFeeValue;
+    }
+  } else {
+    // legacy (and BTCK v1)
+    rawFeeValue = await publicClient.readContract({
+      abi: tokenContract.abi,
+      address: tokenContract.address,
+      functionName: 'getBurnCommission',
+    });
+  }
 
   return fromSatoshi(String(rawFeeValue));
 }