+++ packages/sdk/CHANGELOG.md
@@ -1,12 +1,39 @@
+# 3.6.0
+
+* `Token.LBTC` ABI changes to `stLBTC` for specified chains: `Sepolia`, ...
+* added new `Token.NativeLBTC`,
+* changed `getLBTCMintingFee` and `getMintingFee` so it either takes the value
+  from the token contract (old version) or `AssetRouter` contract (new version),
+* renamed `getBurningFee` to `getRedeemFee` and refactored it so it takes the
+  fee values from the `AssetRouter` or the token contract (old version).
+
+# 3.5.12
+
+* added `tac` chain and bridge eth - tac bridge (OFT).
+
+# 3.5.11
+
+* updated the BFF API urls.
+
+# 3.5.10
+
+* added `token_address` param to generate and get deposit address functions.
+
+# 3.5.9
+
+* changed LBTC addresses for `dev` env.
+
 # 3.5.8
 
 * fixed exports for `getExchangeRatio`
 
 # 3.5.7
 
-* added `getRewardsInfo` function that retrieves the information about earned rewards (yield),
+* added `getRewardsInfo` function that retrieves the information about earned
+  rewards (yield),
 * added `apr` to the `getLBTCStats` function,
-* added `getExchangeRatio` function that gets the exchange ratios of LBTC:BTC and BTC:LBTC.
+* added `getExchangeRatio` function that gets the exchange ratios of LBTC:BTC
+  and BTC:LBTC.
 
 # 3.5.6